requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
azure-storage-blob
opencensus-ext-flask
opencensus-ext-azure
